open-empi-client/pyproject.toml
[tool.poetry]
name = "open-empi-client"
version = "1.0"
description = "A client library for accessing OpenEMPI"

authors = ['James Glenister']

readme = "README.md"
packages = [
    {include = "open_empi_client"},
]
include = ["CHANGELOG.md", "open_empi_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.7"
httpx = ">=0.15.4,<0.24.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py37', 'py38', 'py39']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
